Changed space to short

+++ huffman/onDecompressUtil.c
@@ -90,7 +90,7 @@ int getTrash(byte firstByte) {
 }
 
 int retrieveTreeSize(byte firstByte, byte secondByte) {
-    int size = firstByte & 0b00011111; //getting which bits of the 5 last ones are used
+    short int size = firstByte & 0b00011111; //getting which bits of the 5 last ones are used
     size = size << 8; //left shifting found bits 
     size |= secondByte; //comparing bits of second byte to see what should use
     return size;
